Move provenance components into own dir
