+++ context/app/static/js/components/provenance/ProvGraph/ProvGraph.jsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { renderProvVis } from '../../ProvVis';
+
+import SectionItem from 'js/components/Detail/SectionItem';
+import { renderProvVis } from '../ProvVis';
 import { StyledTypography, StyledLink, FlexPaper } from './style';
-import SectionItem from '../SectionItem';
 import '@hms-dbmi-bgm/react-workflow-viz/dist/react-workflow-viz.min.css';
 
 function ProvGraph(props) {
