+++ context/app/static/js/components/provenance/ProvVis/__tests__/fixtures/real-cwl.json
