+++ context/app/static/js/components/provenance/ProvVis/__tests__/fixtures/simple-cwl.json
